+++ yash-syntax/src/syntax.rs
@@ -19,13 +19,6 @@
 //! This module contains types that represent abstract syntax trees (ASTs) of
 //! the shell language.
 //!
-//! Some types in this module has the type parameter `<H = HereDoc>`. As a user
-//! of ASTs, you will never have to specify the parameter other than the default
-//! `HereDoc`. The parameter with non-default types is used internally by the
-//! parser to create intermediate ASTs that lack sub-trees for here-documents,
-//! as the contents of here-documents have to be parsed separately from the
-//! normal flow of source code.
-//!
 //! ## Syntactic elements
 //!
 //! The AST type that represents the whole shell script is [`List`], which is a
@@ -868,15 +861,15 @@ impl fmt::Display for HereDoc {
 
 /// Part of a redirection that defines the nature of the resulting file descriptor.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub enum RedirBody<H = HereDoc> {
+pub enum RedirBody {
     /// Normal redirection.
     Normal { operator: RedirOp, operand: Word },
     /// Here-document.
-    HereDoc(Rc<H>),
+    HereDoc(Rc<HereDoc>),
     // TODO process redirection
 }
 
-impl RedirBody<HereDoc> {
+impl RedirBody {
     /// Returns the operand word of the redirection.
     pub fn operand(&self) -> &Word {
         match self {
@@ -886,7 +879,7 @@ impl RedirBody<HereDoc> {
     }
 }
 
-impl<H: fmt::Display> fmt::Display for RedirBody<H> {
+impl fmt::Display for RedirBody {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             RedirBody::Normal { operator, operand } => write!(f, "{}{}", operator, operand),
@@ -903,14 +896,14 @@ impl<T: Into<Rc<HereDoc>>> From<T> for RedirBody {
 
 /// Redirection.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct Redir<H = HereDoc> {
+pub struct Redir {
     /// File descriptor that is modified by this redirection.
     pub fd: Option<Fd>,
     /// Nature of the resulting file descriptor.
-    pub body: RedirBody<H>,
+    pub body: RedirBody,
 }
 
-impl<H> Redir<H> {
+impl Redir {
     /// Computes the file descriptor that is modified by this redirection.
     ///
     /// If `self.fd` is `Some(_)`, the `RawFd` value is returned intact. Otherwise,
@@ -927,7 +920,7 @@ impl<H> Redir<H> {
     }
 }
 
-impl<H: fmt::Display> fmt::Display for Redir<H> {
+impl fmt::Display for Redir {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         if let Some(fd) = self.fd {
             write!(f, "{}", fd)?;
@@ -941,13 +934,13 @@ impl<H: fmt::Display> fmt::Display for Redir<H> {
 /// In the shell language syntax, a valid simple command must contain at least one of assignments,
 /// redirections, and words. The parser must not produce a completely empty simple command.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct SimpleCommand<H = HereDoc> {
+pub struct SimpleCommand {
     pub assigns: Vec<Assign>,
     pub words: Vec<Word>,
-    pub redirs: Rc<Vec<Redir<H>>>,
+    pub redirs: Rc<Vec<Redir>>,
 }
 
-impl<H> SimpleCommand<H> {
+impl SimpleCommand {
     /// Returns true if the simple command does not contain any assignments,
     /// words, or redirections.
     pub fn is_empty(&self) -> bool {
@@ -960,7 +953,7 @@ impl<H> SimpleCommand<H> {
     }
 }
 
-impl<H: fmt::Display> fmt::Display for SimpleCommand<H> {
+impl fmt::Display for SimpleCommand {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let i1 = self.assigns.iter().map(|x| x as &dyn fmt::Display);
         let i2 = self.words.iter().map(|x| x as &dyn fmt::Display);
@@ -972,12 +965,12 @@ impl<H: fmt::Display> fmt::Display for SimpleCommand<H> {
 
 /// `elif-then` clause.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ElifThen<H = HereDoc> {
-    pub condition: List<H>,
-    pub body: List<H>,
+pub struct ElifThen {
+    pub condition: List,
+    pub body: List,
 }
 
-impl<H: fmt::Display> fmt::Display for ElifThen<H> {
+impl fmt::Display for ElifThen {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "elif {:#} then ", self.condition)?;
         if f.alternate() {
@@ -990,17 +983,17 @@ impl<H: fmt::Display> fmt::Display for ElifThen<H> {
 
 /// Branch item of a `case` compound command.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct CaseItem<H = HereDoc> {
+pub struct CaseItem {
     /// Array of patterns that are matched against the main word of the case
     /// compound command to decide if the body of this item should be executed.
     ///
     /// A syntactically valid case item must have at least one pattern.
     pub patterns: Vec<Word>,
     /// Commands that are executed if any of the patterns matched.
-    pub body: List<H>,
+    pub body: List,
 }
 
-impl<H: fmt::Display> fmt::Display for CaseItem<H> {
+impl fmt::Display for CaseItem {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
             f,
@@ -1013,37 +1006,34 @@ impl<H: fmt::Display> fmt::Display for CaseItem<H> {
 
 /// Command that contains other commands.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub enum CompoundCommand<H = HereDoc> {
+pub enum CompoundCommand {
     /// List as a command.
-    Grouping(List<H>),
+    Grouping(List),
     /// Command for executing commands in a subshell.
-    Subshell(List<H>),
+    Subshell(List),
     /// For loop.
     For {
         name: Word,
         values: Option<Vec<Word>>,
-        body: List<H>,
+        body: List,
     },
     /// While loop.
-    While { condition: List<H>, body: List<H> },
+    While { condition: List, body: List },
     /// Until loop.
-    Until { condition: List<H>, body: List<H> },
+    Until { condition: List, body: List },
     /// If conditional construct.
     If {
-        condition: List<H>,
-        body: List<H>,
-        elifs: Vec<ElifThen<H>>,
-        r#else: Option<List<H>>,
+        condition: List,
+        body: List,
+        elifs: Vec<ElifThen>,
+        r#else: Option<List>,
     },
     /// Case conditional construct.
-    Case {
-        subject: Word,
-        items: Vec<CaseItem<H>>,
-    },
+    Case { subject: Word, items: Vec<CaseItem> },
     // TODO [[ ]]
 }
 
-impl<H: fmt::Display> fmt::Display for CompoundCommand<H> {
+impl fmt::Display for CompoundCommand {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         use CompoundCommand::*;
         match self {
@@ -1090,14 +1080,14 @@ impl<H: fmt::Display> fmt::Display for CompoundCommand<H> {
 
 /// Compound command with redirections.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct FullCompoundCommand<H = HereDoc> {
+pub struct FullCompoundCommand {
     /// The main part.
-    pub command: CompoundCommand<H>,
+    pub command: CompoundCommand,
     /// Redirections.
-    pub redirs: Vec<Redir<H>>,
+    pub redirs: Vec<Redir>,
 }
 
-impl<H: fmt::Display> fmt::Display for FullCompoundCommand<H> {
+impl fmt::Display for FullCompoundCommand {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let FullCompoundCommand { command, redirs } = self;
         write!(f, "{}", command)?;
@@ -1107,16 +1097,16 @@ impl<H: fmt::Display> fmt::Display for FullCompoundCommand<H> {
 
 /// Function definition command.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct FunctionDefinition<H = HereDoc> {
+pub struct FunctionDefinition {
     /// Whether the function definition command starts with the `function` reserved word.
     pub has_keyword: bool,
     /// Function name.
     pub name: Word,
     /// Function body.
-    pub body: Rc<FullCompoundCommand<H>>,
+    pub body: Rc<FullCompoundCommand>,
 }
 
-impl<H: fmt::Display> fmt::Display for FunctionDefinition<H> {
+impl fmt::Display for FunctionDefinition {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         if self.has_keyword {
             f.write_str("function ")?;
@@ -1127,16 +1117,16 @@ impl<H: fmt::Display> fmt::Display for FunctionDefinition<H> {
 
 /// Element of a pipe sequence.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub enum Command<H = HereDoc> {
+pub enum Command {
     /// Simple command.
-    Simple(SimpleCommand<H>),
+    Simple(SimpleCommand),
     /// Compound command.
-    Compound(FullCompoundCommand<H>),
+    Compound(FullCompoundCommand),
     /// Function definition command.
-    Function(FunctionDefinition<H>),
+    Function(FunctionDefinition),
 }
 
-impl<H: fmt::Display> fmt::Display for Command<H> {
+impl fmt::Display for Command {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Command::Simple(c) => c.fmt(f),
@@ -1148,19 +1138,19 @@ impl<H: fmt::Display> fmt::Display for Command<H> {
 
 /// Commands separated by `|`
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct Pipeline<H = HereDoc> {
+pub struct Pipeline {
     /// Elements of the pipeline.
     ///
     /// A valid pipeline must have at least one command.
     ///
     /// The commands are contained in `Rc` to allow executing them
     /// asynchronously without cloning them.
-    pub commands: Vec<Rc<Command<H>>>,
+    pub commands: Vec<Rc<Command>>,
     /// True if the pipeline begins with a `!`.
     pub negation: bool,
 }
 
-impl<H: fmt::Display> fmt::Display for Pipeline<H> {
+impl fmt::Display for Pipeline {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> fmt::Result {
         if self.negation {
             write!(f, "! ")?;
@@ -1209,12 +1199,12 @@ impl fmt::Display for AndOr {
 
 /// Pipelines separated by `&&` and `||`.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct AndOrList<H = HereDoc> {
-    pub first: Pipeline<H>,
-    pub rest: Vec<(AndOr, Pipeline<H>)>,
+pub struct AndOrList {
+    pub first: Pipeline,
+    pub rest: Vec<(AndOr, Pipeline)>,
 }
 
-impl<H: fmt::Display> fmt::Display for AndOrList<H> {
+impl fmt::Display for AndOrList {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self.first)?;
         self.rest
@@ -1225,12 +1215,12 @@ impl<H: fmt::Display> fmt::Display for AndOrList<H> {
 
 /// Element of a [List].
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct Item<H = HereDoc> {
+pub struct Item {
     /// Main part of this item.
     ///
     /// The and-or list is contained in `Rc` to allow executing it
     /// asynchronously without cloning it.
-    pub and_or: Rc<AndOrList<H>>,
+    pub and_or: Rc<AndOrList>,
     /// Location of the `&` operator for this item, if any.
     pub async_flag: Option<Location>,
 }
@@ -1240,7 +1230,7 @@ pub struct Item<H = HereDoc> {
 /// By default, the `;` terminator is omitted from the formatted string.
 /// When the alternate flag is specified as in `{:#}`, the result is always
 /// terminated by either `;` or `&`.
-impl<H: fmt::Display> fmt::Display for Item<H> {
+impl fmt::Display for Item {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self.and_or)?;
         if self.async_flag.is_some() {
@@ -1257,14 +1247,14 @@ impl<H: fmt::Display> fmt::Display for Item<H> {
 ///
 /// It depends on context whether an empty list is a valid syntax.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct List<H = HereDoc>(pub Vec<Item<H>>);
+pub struct List(pub Vec<Item>);
 
 /// Allows conversion from List to String.
 ///
 /// By default, the last `;` terminator is omitted from the formatted string.
 /// When the alternate flag is specified as in `{:#}`, the result is always
 /// terminated by either `;` or `&`.
-impl<H: fmt::Display> fmt::Display for List<H> {
+impl fmt::Display for List {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         if let Some((last, others)) = self.0.split_last() {
             for item in others {